pyproject.toml
[project]
name = "mlx-8-backprop-week-1-hackernews-vote-predictions"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas>=2.3.0",
    "scipy>=1.15.3",
    "torch>=2.7.1",
    "transformers>=4.52.4",
    "psycopg2-binary>=2.9.0",
    "scikit-learn>=1.3.0",
]

# https://docs.astral.sh/uv/concepts/projects/workspaces/
[tool.uv.workspace]
members = ["packages/*"]
